Remove POD/Author tests
This is a little unusual because they've subclassed Module::Build
to so some horrible things that mean there are *2* manifests, one for tests
which it also pointlessly tests.

So stock deleting the bad tests results in test failures.

The only way to subjugate this bad behaviour is patch out the use of the
test-specific sub-class to avoid those pointless testing shenanigans.